Fix the video unable to pause after MoviePlayer resumed
When video is playing, relaunch of MoviePlayer can be triggered by
density change. In this case, it will resume from savedInstance.
MoviePlayer wants to pretend that the video has started by calling
mVideoView.start(). But it forgot to update the state of MovieController,
thus the controller is still in LOADING state even the video has resumed
playing. Due to the mismatch, the controller UI stop working.

Change-Id: Ida34c08765d75f8f68a20c7ce39555a091a9bd57
Tracked-On:
Signed-off-by: Lei,RayX <rayx.lei@intel.com>